Improve boo--multi-line->single-line and also remove a debug message that slipped through the crack
Before it would fail if the conditional was the first line of the file
since it would delete backwards in order to move the body of the
conditional up. It would then de-indent the body of the conditional.
Now it will instead save the current-column, delet character forward
and indent the line to that column which will put it in the
appropriate place.